+++ packages/expo-dev-menu/android/src/debug/java/expo/modules/devmenu/DevMenuDevSettings.kt
@@ -1,6 +1,7 @@
 package expo.modules.devmenu
 
 import com.facebook.react.ReactHost
+import expo.modules.devmenu.api.CustomPerformanceMonitor
 import expo.modules.devmenu.devtools.DevMenuDevToolsDelegate
 import expo.modules.kotlin.weak
 
@@ -12,13 +13,16 @@ data class DevToolsSettings(
 
 object DevMenuDevSettings {
   fun getDevSettings(reactHost: ReactHost): DevToolsSettings {
-    val devDelegate = DevMenuDevToolsDelegate(requireNotNull(reactHost.devSupportManager).weak())
+    val devSupportManager = requireNotNull(reactHost.devSupportManager)
+    val devDelegate = DevMenuDevToolsDelegate(devSupportManager.weak())
     val devSettings = devDelegate.devSettings
 
     return DevToolsSettings(
       isElementInspectorShown = devSettings?.isElementInspectorEnabled ?: false,
       isHotLoadingEnabled = devSettings?.isHotModuleReplacementEnabled ?: true,
-      isPerfMonitorShown = devSettings?.isFpsDebugEnabled ?: false
+      isPerfMonitorShown = (devSupportManager as? CustomPerformanceMonitor)?.isPerformanceMonitorShown
+        ?: devSettings?.isFpsDebugEnabled
+        ?: false
     )
   }
 }