+++ packages/expo-router/src/native-tabs/types.ts
@@ -498,6 +498,7 @@ export interface NativeTabsProps extends PropsWithChildren {
 
 export interface InternalNativeTabsProps extends NativeTabsProps {
   nonTriggerChildren?: React.ReactNode;
+  tabConfigurationKey: string;
 }
 export interface OnTabChangeEventPayload {
   /**
@@ -536,6 +537,7 @@ export interface NativeTabsViewProps extends Omit<
   | 'rippleColor'
   | 'disableIndicator'
   | 'labelVisibilityMode'
+  | 'tabConfigurationKey'
 > {
   focusedIndex: number;
   /**